requirements.txt
flask
weasyprint
google-genai
python-dotenv
